pyproject.toml
[project]
name = "ctfbot"
version = "0.1.0"
description = "discord bot for CTF"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "ctftime-api>=0.1.5",
    "discord-py>=2.4.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
    "ty>=0.0.1a8",
]

[tool.ruff]
# Target Python version
target-version = "py312"

# Line length and indentation
line-length = 88
indent-width = 4

# Exclude directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
select = ["E", "F", "I", "W", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ty]
# Source root configuration
[tool.ty.src]
root = "./src"

# Environment configuration
[tool.ty.environment]
python = "./.venv"
python-version = "3.12"
